+++ include/shark/Algorithms/Trainers/LassoRegression.h
@@ -213,7 +213,7 @@ protected:
 				if ((double)rand() / (double)RAND_MAX < prob) m++;
 				for (std::size_t  j=0; j<m; j++)
 				{
-					index[pos] = i;
+					index[pos] = (unsigned int)i;
 					pos++;
 				}
 				psum -= p;
