+++ include/shark/LinAlg/KernelMatrix.h
@@ -109,7 +109,7 @@ public:
         m_accessCounter += end-start;
         
         typename AbstractKernelFunction<InputType>::ConstInputReference xi = *x[i];
-        SHARK_PARALLEL_FOR(int j = start; j < (int) end; j++)
+        SHARK_PARALLEL_FOR(int j = (int)start; j < (int) end; j++)
         {
             storage[j-start] = QpFloatType(kernel.eval(xi, *x[j]));
         }
